Create bundle:cleanup rake task to clean out dummy data
In order to start with a clean slate after importing the bundle, we need to delete a number of existing dummy records created by the import process. This new rake task will allow us to start fresh after either the first install (Pivotal Chore #110669210), or even between imports of our own data in dev if we so choose (Pivotal Chore #112362835).

Also, rename the previous `bundle:reset` task to `bundle:uninstall` to be the counter-weight to the `bundle:download_and_install`. Use `bundle:cleanup` as part of `bundle:uninstall` to avoid duplication.