Facilitate Cypress external validation by referencing HQMF in global namespace
In order to allow projects to use Cypress for external validation, we need to change some of the ways that we reference some of our internal namespaces. One such namespace is HQMF which is generated during `lib/hqmf-parser.rb`. In order to allow others to use this module as well from outside this project, reference it in the global namespace.