Return Record object after importing cat1 XML file
In order to be able to use the newly-created record file from outside the bulk importer WITHOUT introducing a race condition with Record.last, return the newly-created record object from the import method. Note: this used to return true, so whoever is using it can still expect a truthy check to be valid (e.g., if import); however, any specific checks (e.g., if import == true) will no longer work. Since we are the only ones using this at the moment, however, this is acceptable, and for sure perferable to the other hoops we were jumping through in order to find the newly-created Record object.