Add versioning to bundle uninstall
When uninstalling we may want to uninstall a single version of a bundle.

We do remove the call to `Rake::Task['bundle:cleanup'].execute` because
this removes all patient and query caches. These are not bundle
specific and cannot be safely removed. However we are not importing
sample patients by default anymore so the artifacts from a bundle import
should not exist anyways.